64bits fixes.
On 64bits platforms, ssize_t is an int64_t.

Given the average string length pfixtools will handle, a 32bits integer is
enough, let's use int everywhere.

Also a pointer difference is an ptrdiff_t which is also a 64bits integer, so
when used as the width specifier of a %.*s format, it must be cast into an
int.

Signed-off-by: Pierre Habouzit <madcoder@debian.org>